% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_percent_growth.R
\name{calc_percent_growth}
\alias{calc_percent_growth}
\title{Calculates percentage growth on an absolute or per period average basis}
\usage{
calc_percent_growth(
  data,
  period_col,
  start_period,
  end_period,
  periods,
  value_col,
  calculation,
  ...
)
}
\arguments{
\item{data}{Dataframe holding the data you would like to perform the calculation for.}

\item{period_col}{The column containing data for the period the value you are calculating pertains to.}

\item{start_period}{The first period you want to use for your calculations. Must be present in the \code{period_col}.}

\item{end_period}{The last period you want to use for your calculations. Must be present in the \code{period_col}.}

\item{periods}{Number of periods covered in the calculation. First period is period 0.}

\item{value_col}{The column containing the value you want to calculate percentage growth for.}

\item{calculation}{Specify whether you would like to calculate growth on an absolute or average per period basis.}

\item{...}{Grouping variables if the data needs to be aggregated.}
}
\value{
A dataframe containing the calculated percentages for the specified grouping variables and periods considered. It will be listed for the start period to the end period.
}
\description{
Allows the calculation of percentage change in a variable over user-defined periods of time, on either an absolute basis, or averaged over the course of the period.
}
