% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weighted_mean.R
\name{calc_weighted_mean}
\alias{calc_weighted_mean}
\title{Calculates a weighted mean based on grouping criteria you set}
\usage{
calc_weighted_mean(data, numerator_col, denominator_col, multiplier = 1, ...)
}
\arguments{
\item{data}{The data frame/table holding the information you'd like to calculate a weighted mean for}

\item{numerator_col}{Value field that when aggregated will be the numerator in the calculation to work out weighted mean}

\item{denominator_col}{Value field that when aggregated will be the denominator in the calculation to work out a weighted mean}

\item{multiplier}{A figure to divide your end result by if your prior processing involves counting your denominator multiple times}

\item{...}{Any columns you would like to group the data by when aggregating}
}
\value{
A data frame containing the weighted mean for each group of data calculated
}
\description{
Takes a dataframe and calculates a weighted mean based on a chosen denominator and numerator column.
}
