% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_over_time.R
\name{profile_over_time}
\alias{profile_over_time}
\title{Create a 'straight-line' profile of a value over its associated lifetime, split up into user-defined periods.}
\usage{
profile_over_time(
  df,
  begin_col,
  end_col,
  period_start,
  period_end,
  period_breaks,
  value_col,
  ...
)
}
\arguments{
\item{df}{Dataframe holding the data you would like to perform the calculation for.}

\item{begin_col}{Column containing information about the begin date for the item you're interested in. Must be in date format. Will be renamed 'start_date' during processing.}

\item{end_col}{Column ontaining the information about the end date of the item you're interested in. Must be in date format. Will be renamed 'end_date' during processing.}

\item{period_start}{The beginning of the period you would like to profile over. Enter as "yyyy-mm-dd".}

\item{period_end}{The end of the period you would like to profile over. Enter as "yyyy-mm-dd".}

\item{period_breaks}{Defines the length of each period. Can be "day", "week", "month", "quarter" or "year".}

\item{value_col}{The value you'd like to profile. Must be numeric.}

\item{...}{Columns containing information about the item the value is attached to.}
}
\value{
A dataframe containing the data profiled over the specified periods.
}
\description{
Can take a value associated with an item and spreads the value over the item's lifetime. One example would be profiling income/expenditure using the total value of a five year contract.
}
